backend/requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
openai
python-multipart
psycopg2-binary
pydantic